Refactor ProductDisplay
