+++ Source/LibationWinForms/AvaloniaUI/Views/MainWindow/MainWindow.axaml
@@ -5,7 +5,6 @@
 		xmlns:vm="clr-namespace:LibationWinForms.AvaloniaUI.ViewModels"
 		xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 		xmlns:views="clr-namespace:LibationWinForms.AvaloniaUI.Views"
-		xmlns:prgid="clr-namespace:LibationWinForms.AvaloniaUI.Views.ProductsGrid"
 		xmlns:controls="clr-namespace:LibationWinForms.AvaloniaUI.Controls"
 		mc:Ignorable="d" d:DesignWidth="1850" d:DesignHeight="700"
 		x:Class="LibationWinForms.AvaloniaUI.Views.MainWindow"
@@ -172,7 +171,7 @@
 					</SplitView.Pane>
 
 					<!-- Product Display Grid -->
-					<prgid:ProductsDisplay2
+					<views:ProductsDisplay2
 						DataContext="{Binding ProductsDisplay}"						
 						LiberateClicked="ProductsDisplay_LiberateClicked"
 						Name="productsDisplay"  />
